substrait::df_logical_plan

Module consumer

Structs

DefaultSubstraitConsumer

Default SubstraitConsumer for converting standard Substrait without user-defined extensions.

ExprContainer

An ExprContainer is a container for a collection of expressions with a common input schema

Traits

SubstraitConsumer

This trait is used to consume Substrait plans, converting them into DataFusion Logical Plans. It can be implemented by users to allow for custom handling of relations, expressions, etc.

Functions

apply_masking

from_aggregate_rel

from_cast

from_cross_rel

from_exchange_rel

from_fetch_rel

from_field_reference

from_filter_rel

from_if_then

from_join_rel

from_literal

from_project_rel

from_read_rel

from_scalar_function

from_set_rel

from_singular_or_list

from_sort_rel

from_subquery

from_substrait_agg_func

Convert Substrait AggregateFunction to DataFusion Expr

from_substrait_extended_expr

Convert Substrait ExtendedExpression to ExprContainer

from_substrait_func_args

Convert Substrait FunctionArguments to DataFusion Exprs

from_substrait_named_struct

Convert Substrait NamedStruct to DataFusion DFSchemaRef

from_substrait_plan

Convert Substrait Plan to DataFusion LogicalPlan

from_substrait_plan_with_consumer

Convert Substrait Plan to DataFusion LogicalPlan using the given consumer

from_substrait_rel

Convert Substrait Rel to DataFusion DataFrame

from_substrait_rex

Convert Substrait Rex to DataFusion Expr

from_substrait_rex_vec

Convert Substrait Expressions to DataFusion Exprs

from_substrait_sorts

Convert Substrait Sorts to DataFusion Exprs

from_window_function

name_to_op

substrait_fun_name