substrait::df_logical_plan

Module producer

Traits

SubstraitProducer

This trait is used to produce Substrait plans, converting them from DataFusion Logical Plans. It can be implemented by users to allow for custom handling of relations, expressions, etc.

Functions

from_aggregate

from_aggregate_function

from_alias

from_between

from_binary_expr

from_case

from_cast

from_column

from_distinct

from_empty_relation

from_filter

from_in_list

from_in_subquery

from_join

from_like

from_limit

from_literal

from_projection

from_repartition

from_scalar_function

from_sort

from_subquery_alias

from_table_scan

from_try_cast

from_unary_expr

from_union

from_values

from_window

from_window_function

make_binary_op_scalar_func

Return Substrait scalar function with two arguments

operator_to_name

parse_flat_grouping_exprs

to_substrait_agg_measure

to_substrait_extended_expr

Serializes a collection of expressions to a Substrait ExtendedExpression message

to_substrait_groupings

to_substrait_plan

Convert DataFusion LogicalPlan to Substrait Plan

to_substrait_rel

to_substrait_rex

Convert DataFusion Expr to Substrait Rex