substrait

Module variation_const

Type variation constants

To add support for types not in the core specification, we make use of the simple extensions of substrait type. This module contains the constants used to identify the type variation.

The rules of type variations here are:

Default type reference is 0. It is used when the actual type is the same with the original type.

Extended variant type references start from 1, and usually increase by 1.

TODO: Definitions here are not the final form. All the non-system-preferred variations will be defined using simple extensions as per the spec of type_variations https://github.com/apache/datafusion/issues/11545

Constants

DATE_32_TYPE_VARIATION_REF

DATE_64_TYPE_VARIATION_REF

DECIMAL_128_TYPE_VARIATION_REF

DECIMAL_256_TYPE_VARIATION_REF

DEFAULT_CONTAINER_TYPE_VARIATION_REF

DEFAULT_TYPE_VARIATION_REF

The “system-preferred” variation (i.e., no variation).

DICTIONARY_TYPE_VARIATION_REF

INTERVAL_DAY_TIME_TYPE_REFDeprecated

For DataType::Interval with IntervalUnit::DayTime.

INTERVAL_MONTH_DAY_NANO_TYPE_NAMEDeprecated

For DataType::Interval with IntervalUnit::MonthDayNano.

INTERVAL_MONTH_DAY_NANO_TYPE_REFDeprecated

For DataType::Interval with IntervalUnit::MonthDayNano.

INTERVAL_YEAR_MONTH_TYPE_REFDeprecated

For DataType::Interval with IntervalUnit::YearMonth.

LARGE_CONTAINER_TYPE_VARIATION_REF

TIMESTAMP_MICRO_TYPE_VARIATION_REFDeprecated

TIMESTAMP_MILLI_TYPE_VARIATION_REFDeprecated

TIMESTAMP_NANO_TYPE_VARIATION_REFDeprecated

TIMESTAMP_SECOND_TYPE_VARIATION_REFDeprecated

UNSIGNED_INTEGER_TYPE_VARIATION_REF

VIEW_CONTAINER_TYPE_VARIATION_REF